project2/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyWavelets>=1.5.0
scikit-image>=0.22.0
matplotlib>=3.7.0
